# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "篮球课堂训练复盘",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analyses/upload": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Upload Analysis",
        "operationId": "upload_analysis_api_v1_analyses_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_analysis_api_v1_analyses_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analyses/{analysis_id}": {
      "get": {
        "tags": [
          "analyses"
        ],
        "summary": "Get Analysis",
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "analyses"
        ],
        "summary": "Delete Analysis",
        "operationId": "delete_analysis_api_v1_analyses__analysis_id__delete",
        "parameters": [
          {
            "name": "analysis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Analysis Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks": {
      "post": {
        "tags": [
          "tasks"
        ],
        "summary": "Create Task",
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "List Tasks",
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 120
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "quick",
                    "full"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mode"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "uploading",
                    "queued",
                    "running",
                    "completed",
                    "failed",
                    "canceled",
                    "expired"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskListPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tasks/{task_id}": {
      "patch": {
        "tags": [
          "tasks"
        ],
        "summary": "Update Task",
        "operationId": "update_task_api_v1_tasks__task_id__patch",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "tasks"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPublic"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "tasks"
        ],
        "summary": "Delete Task",
        "operationId": "delete_task_api_v1_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalysisPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale",
            "default": "zh"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode",
            "default": "sequential"
          },
          "expected_persons": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Persons"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "stage_message": {
            "type": "string",
            "title": "Stage Message"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "owner_id": {
            "type": "integer",
            "title": "Owner Id"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "mode",
          "source_type",
          "preset_id",
          "status",
          "progress",
          "stage_message",
          "error_code",
          "error_message",
          "owner_id",
          "submitted_at",
          "created_via",
          "retry_count",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at"
        ],
        "title": "AnalysisPublic"
      },
...
      "Body_upload_analysis_api_v1_analyses_upload_post": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode"
          },
          "expected_persons": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 1.0,
            "title": "Expected Persons"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale"
          },
          "sync": {
            "type": "string",
            "maxLength": 8192,
            "title": "Sync"
          },
          "enrollment_video": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Enrollment Video"
          },
          "cam_01": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 01"
          },
          "cam_02": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 02"
          },
          "cam_03": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 03"
          },
          "cam_04": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Cam 04"
          }
        },
        "type": "object",
        "required": [
          "title",
          "mode",
          "enrollment_mode",
          "expected_persons",
          "analyst_locale",
          "sync",
          "enrollment_video",
          "cam_01",
          "cam_02",
          "cam_03",
          "cam_04"
        ],
        "title": "Body_upload_analysis_api_v1_analyses_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode",
            "default": "full"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale",
            "default": "zh"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode",
            "default": "sequential"
          },
          "expected_persons": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Persons"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "TaskCreate"
      },
...
      "TaskListPublic": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TaskPublic"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "TaskListPublic"
      },
      "TaskPublic": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode"
          },
          "analyst_locale": {
            "type": "string",
            "enum": [
              "zh",
              "en"
            ],
            "title": "Analyst Locale",
            "default": "zh"
          },
          "enrollment_mode": {
            "type": "string",
            "enum": [
              "sequential",
              "lineup"
            ],
            "title": "Enrollment Mode",
            "default": "sequential"
          },
          "expected_persons": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Persons"
          },
          "sync_status": {
            "type": "string",
            "enum": [
              "unconfirmed",
              "confirmed",
              "stale",
              "legacy"
            ],
            "title": "Sync Status",
            "default": "unconfirmed"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preset Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "uploading",
              "queued",
              "running",
              "completed",
              "failed",
              "canceled",
              "expired"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress"
          },
          "stage_message": {
            "type": "string",
            "title": "Stage Message"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "created_via": {
            "type": "string",
            "title": "Created Via"
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/TaskInputPublic"
            },
            "type": "array",
            "title": "Inputs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "mode",
          "source_type",
          "preset_id",
          "status",
          "progress",
          "stage_message",
          "error_code",
          "error_message",
          "submitted_at",
          "created_via",
          "retry_count",
          "created_at",
          "updated_at",
          "started_at",
          "completed_at",
          "inputs"
        ],
        "title": "TaskPublic"
      },
      "TaskUpdate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Title"
          },
          "mode": {
            "type": "string",
            "enum": [
              "quick",
              "full"
            ],
            "title": "Mode"
          },
          "analyst_locale": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "zh",
                  "en"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyst Locale"
          },
          "enrollment_mode": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "sequential",
                  "lineup"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Enrollment Mode"
          },
          "expected_persons": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 6.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Persons"
          }
        },
        "type": "object",
        "required": [
          "title",
          "mode"
        ],
        "title": "TaskUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}